relay-trainer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relay-trainer"
version = "0.1.0"
description = "Relay training control plane and worker runtime for Lium pods"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "requests>=2.32.0",
  "typer>=0.12.0",
  "huggingface_hub>=0.34.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["relay*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
